refactor: create own view and separate data file
