+++ src/views/well-architected-framework/well-architected-framework-landing.module.css
@@ -1,0 +1,6 @@
+.hero {
+	margin-bottom: 38px;
+}
+.overview {
+	margin-bottom: 48px;
+}
